rest-api/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==2.2.5
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
docopt==0.6.2
idna==2.8
mysqlclient==1.4.5
pipreqs==0.4.10
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
yarg==0.1.9
